requirements.txt
# Koicha SEO Engine — Requirements
# Python 3.11+

# PDF Parsing
pdfplumber==0.10.3

# Google AI
google-generativeai==0.5.4

# Google APIs
google-auth==2.28.1
google-auth-oauthlib==1.2.0
google-auth-httplib2==0.2.0
google-api-python-client==2.120.0

# HTTP / Scraping
requests==2.31.0
beautifulsoup4==4.12.3

# Telegram Bot
python-telegram-bot==20.7

# Instagram scraping (no API key needed)
instaloader==4.10.3

# Scheduling (replaces Cloud Scheduler locally)
schedule==1.2.1

# Templating (for static site generation)
jinja2==3.1.3

# QR Code generation (for review prompt cards)
qrcode[pil]==7.4.2

# Environment variables
python-dotenv==1.0.1

# Image processing
Pillow==10.2.0

# Schema.org generation helper
schemaorg==0.0.1

# Utilities
lxml==5.1.0
cssutils==2.9.0
